requirements.txt
fitz==0.0.1.dev2
moderngl==5.12.0
numpy==1.24.3
Pillow==12.1.1
